chkstat: drop types.h header
With the now clearer split-up responsibilities there's no need for an
assorted types header anymore. Move the remaining ProfileEntry into
profparser.h, where it belongs semantically, and drop the types.h
header.
